refactor: reorganize decode vectorized code
