+++ rust/mlt/src/decoder/helpers.rs
@@ -1,4 +1,5 @@
 use bytes::Buf;
+use zigzag::ZigZag;
 
 use crate::decoder::tracked_bytes::TrackedBytes;
 use crate::metadata::proto_tileset::{Column, ScalarType, column, scalar_column};
@@ -60,6 +61,37 @@ pub fn get_data_type_from_column(column_metadata: &Column) -> MltResult<ScalarTy
     }
 }
 
+/// Decode ([`ZigZag`] + delta) for Vec2s
+// TODO: The encoded process is (delta + ZigZag) for each component
+pub fn decode_componentwise_delta_vec2s<T: ZigZag>(data: &[T::UInt]) -> Result<Vec<T>, MltError> {
+    let len = data.len();
+    if len < 2 {
+        return Err(MltError::MinLength {
+            ctx: "vec2 delta stream",
+            min: 2,
+            got: len,
+        });
+    }
+    if len % 2 != 0 {
+        return Err(MltError::InvalidValueMultiple {
+            ctx: "vec2 delta stream length",
+            multiple_of: 2,
+            got: len,
+        });
+    }
+
+    let mut result = Vec::with_capacity(len);
+    result.push(T::decode(data[0]));
+    result.push(T::decode(data[1]));
+
+    for i in (2..len).step_by(2) {
+        result.push(T::decode(data[i]) + result[i - 2]);
+        result.push(T::decode(data[i + 1]) + result[i - 1]);
+    }
+
+    Ok(result)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -92,5 +124,22 @@ mod tests {
         let data_type =
             get_data_type_from_column(&column_metadata).expect("should parse ScalarType");
         assert_eq!(data_type, ScalarType::Uint32);
+    }
+
+    #[test]
+    fn test_decode_componentwise_delta_vec2s() {
+        // original Vec2s: [(3, 5), (7, 6), (12, 4)]
+        // delta:          [3, 5, 4, 1, 5, -2]
+        // ZigZag:         [6, 10, 8, 2, 10, 3]
+        let encoded_from_positives: Vec<u32> = vec![6, 10, 8, 2, 10, 3];
+        let decoded = decode_componentwise_delta_vec2s::<i32>(&encoded_from_positives).unwrap();
+        assert_eq!(decoded, vec![3, 5, 7, 6, 12, 4]);
+
+        // original Vec2s: [(3, 5), (-1, 6), (4, -4)]
+        // delta:          [3, 5, -4, 1, 5, -10]
+        // ZigZag:         [6, 10, 7, 2, 10, 19]
+        let encoded_from_negatives: Vec<u32> = vec![6, 10, 7, 2, 10, 19];
+        let decoded = decode_componentwise_delta_vec2s::<i32>(&encoded_from_negatives).unwrap();
+        assert_eq!(decoded, vec![3, 5, -1, 6, 4, -4]);
     }
 }
