+++ rust/mlt/src/decoder/mod.rs
@@ -1,6 +1,6 @@
 mod decode;
 mod helpers;
 pub mod integer;
+pub mod integer_stream;
 pub mod tracked_bytes;
 pub mod varint;
-pub mod vectorized;
