--- rust/mlt/src/decoder/vectorized/mod.rs
@@ -1,1 +1,0 @@
-pub mod helpers;
